Backend/requirements.txt
google-genai
firebase-admin
